some changes 4
