+++ src/misc.h
@@ -17,6 +17,9 @@
 #define STREQ(x, y) (!strcmp(x, y))
 #define FILE_EXISTS(file) (!access(file, F_OK))
 
+/* two-dimensional arrays only */
+#define ARR_LEN(array) (sizeof(array) / 16)
+
 
 /* Text and logo colorization constants   */
 #define TNRM "\x1B[0;0m"  /* normal       */
@@ -38,58 +41,8 @@
 #define TWHT "\x1B[1;37m" /* white        */
 
 
-/* disp.c and detect.c modules */
-#define ERR_REPORT(str) (fprintf(stderr, TWHT "[[ " TLRD "!" TWHT " ]] " TNRM "Error: %s\n", str))
-#define VERBOSE_OUT(str1, str2) (fprintf(stdout, TLRD ":: " TNRM "%s%s\n", str1, str2))
-
-
-/*  Manifest constants used to control the 'for'
-    loop iteration cycles in some of the disp.c
-    functions, they also appear in structs.h
-*/
-#define MICROSOFT_LEN     16
-#define MACOSX_LEN        16
-#define ARCH_OLD_LEN      18
-#define ARCH_LEN          19
-#define MINT_LEN          18
-#define LMDE_LEN          18
-#define UBUNTU_LEN        18
-#define DEBIAN_LEN        18
-#define CRUNCHBANG_LEN    18
-#define GENTOO_LEN        18
-#define FUNTOO_LEN        18
-#define FEDORA_LEN        18
-#define MANDRAKE_LEN      18
-#define OPENSUSE_LEN      18
-#define SLACKWARE_LEN     21
-#define REDHAT_LEN        18
-#define FRUGALWARE_LEN    23
-#define PEPPERMINT_LEN    18
-#define SOLUSOS_LEN       18
-#define MAGEIA_LEN        18
-#define PARABOLA_LEN      18
-#define VIPERR_LEN        18
-#define DEEPIN_LEN        18
-#define CHAKRA_LEN        18
-#define FUDUNTU_LEN       21
-#define TRISQUEL_LEN      18
-#define MANJARO_LEN       18
-#define ELEMENTARY_LEN    18
-#define SCIENTIFIC_LEN    20
-#define BACKTRACK_LEN     21
-#define KALI_LEN          21
-#define SABAYON_LEN       18
-#define ANDROID_LEN       16
-#define ANGSTROM_LEN      16
-#define LINUX_LEN         18
-#define FREEBSD_LEN       18
-#define OPENBSD_LEN       23
-#define DRAGONFLYBSD_LEN  18
-#define NETBSD_LEN        19
-#define SUNOS_LEN         17
-
-
 /* internationalization macros */
+#include "config.h" /* Auto generated from configure */
 #ifdef _LIBC
 # include <libintl.h>
 # undef dgettext
@@ -99,6 +52,12 @@
 # include "gettext.h"
 #endif
 #define N_(msgid) msgid
+#define _(str) gettext(str)
+
+
+/* feedback */
+#define ERR_REPORT(str) (fprintf(stderr, TWHT "[[ " TLRD "!" TWHT " ]] " TNRM "%s: %s\n", _("Error"), str))
+#define VERBOSE_OUT(str1, str2) (fprintf(stdout, TLRD ":: " TNRM "%s%s\n", str1, str2))
 
 
 #endif /* SCREENFETCH_C_MISC_H */
