+++ src/plat/darwin/detect.c
@@ -1,5 +1,6 @@
 /*	detect.c
  *	Author: William Woodruff
+ *  Edited by: Aaron Caffrey
  *	-------------
  *
  *	The detection functions used by screenfetch-c on OS X (Darwin) are implemented here.
@@ -12,7 +13,7 @@
 #include <stdbool.h>
 #include <string.h>
 #include <unistd.h>
-#include <getopt.h>
+#include <inttypes.h>
 
 /* OS X-specific includes */
 #include <sys/types.h>
@@ -36,7 +37,7 @@
 	detects the computer's distribution (OS X release)
 	argument char *str: the char array to be filled with the distro name
 */
-void detect_distro(char *str)
+void detect_distro(char *str1)
 {
 #if __MAC_OS_X_VERSION_MIN_REQUIRED <= 1070
 	int ver_maj, ver_min, ver_bug;
@@ -53,13 +54,13 @@ void detect_distro(char *str)
 	Gestalt(gestaltSystemVersionMinor, (SInt32 *) &ver_min);
 	Gestalt(gestaltSystemVersionBugFix, (SInt32 *) &ver_bug);
 
-	snprintf(str, MAX_STRLEN, "Max OS X %d.%d.%d", ver_maj, ver_min, ver_bug);
+	snprintf(str1, MAX_STRLEN, "Max OS X %d.%d.%d", ver_maj, ver_min, ver_bug);
 #else
 	distro_file = popen("sw_vers -productVersion | tr -d '\\n'", "r");
 	fgets(distro_name_str, MAX_STRLEN, distro_file);
 	pclose(distro_file);
 
-	snprintf(str, MAX_STRLEN, "Mac OS X %s", distro_name_str);
+	snprintf(str1, MAX_STRLEN, "Mac OS X %s", distro_name_str);
 #endif
 
 	return;
@@ -69,7 +70,7 @@ void detect_distro(char *str)
 	detects the computer's hostname and active user and formats them
 	argument char *str: the char array to be filled with the host info
 */
-void detect_host(char *str)
+void detect_host(void)
 {
 	char *given_user = "Unknown";
 	char given_host[MAX_STRLEN] = "Unknown";
@@ -79,7 +80,8 @@ void detect_host(char *str)
 	uname(&host_info);
 	safe_strncpy(given_host, host_info.nodename, MAX_STRLEN);
 
-	snprintf(str, MAX_STRLEN, "%s@%s", given_user, given_host);
+	snprintf(UseR, MAX_STRLEN, "%s", given_user);
+	snprintf(HosT, MAX_STRLEN, "%s", given_host);
 
 	return;
 }
@@ -105,11 +107,8 @@ void detect_kernel(char *str)
 */
 void detect_uptime(char *str)
 {
-	long long uptime = 0;
-	int secs = 0;
-	int mins = 0;
-	int hrs = 0;
-	int days = 0;
+	long uptime = 0;
+	uint_least32_t secs = 0, mins = 0, hrs = 0, days = 0;
 
 	/* three cheers for undocumented functions and structs */
 	static mach_timebase_info_data_t timebase_info;
@@ -119,16 +118,20 @@ void detect_uptime(char *str)
 		(void) mach_timebase_info(&timebase_info);
 	}
 
-	uptime = (long long)((mach_absolute_time() * timebase_info.numer) /
+	uptime = (long)((mach_absolute_time() * timebase_info.numer) /
 				(1000* 1000 * timebase_info.denom));
 	uptime /= 1000;
 
 	split_uptime(uptime, &secs, &mins, &hrs, &days);
 
 	if (days > 0)
-		snprintf(str, MAX_STRLEN, "%dd %dh %dm %ds", days, hrs, mins, secs);
+		snprintf(str, MAX_STRLEN,
+			"%"PRIuLEAST32"d %"PRIuLEAST32"h %"PRIuLEAST32"m %"PRIuLEAST32"s",
+			days, hrs, mins, secs);
 	else
-		snprintf(str, MAX_STRLEN, "%dh %dm %ds", hrs, mins, secs);
+		snprintf(str, MAX_STRLEN,
+			"%"PRIuLEAST32"h %"PRIuLEAST32"m %"PRIuLEAST32"s",
+			hrs, mins, secs);
 
 	return;
 }
@@ -139,21 +142,18 @@ void detect_uptime(char *str)
 */
 void detect_pkgs(char *str, const char *distro_str)
 {
-	int packages = 0;
+	uint_fast16_t packages = 0;
 	glob_t gl;
 
 	if (glob("/usr/local/Cellar/*", GLOB_NOSORT, NULL, &gl) == 0)
-	{
 		packages = gl.gl_pathc;
-	}
+
 	else if (error)
-	{
-		ERR_REPORT("Failure while globbing packages.");
-	}
+		ERR_REPORT(_("Failure while globbing packages."));
 
 	globfree(&gl);
 
-	snprintf(str, MAX_STRLEN, "%d", packages);
+	snprintf(str, MAX_STRLEN, "%"PRIuFAST16, packages);
 
 	return;
 }
@@ -176,6 +176,8 @@ void detect_cpu(char *str)
 				"tr -d '\\n' | tr -s ' '", "r");
 	fgets(str, MAX_STRLEN, cpu_file);
 	pclose(cpu_file);
+
+	remove_excess_cpu_txt(str);
 
 	return;
 }
@@ -204,7 +206,7 @@ void detect_gpu(char *str)
 void detect_disk(char *str)
 {
 	struct statfs disk_info;
-	long disk_total = 0, disk_used = 0, disk_percentage = 0;
+	uintmax_t disk_total = 0, disk_used = 0, disk_percentage = 0;
 
 	if (!(statfs(getenv("HOME"), &disk_info)))
 	{
@@ -212,13 +214,11 @@ void detect_disk(char *str)
 		disk_used = (((disk_info.f_blocks - disk_info.f_bfree)
 					* disk_info.f_bsize) / GB);
 		disk_percentage = (((float) disk_used / disk_total) * 100);
-		snprintf(str, MAX_STRLEN, "%ldG / %ldG (%ld%%)", disk_used, disk_total,
-				disk_percentage);
+		snprintf(str, MAX_STRLEN, "%"PRIuMAX"G / %"PRIuMAX"G (%"PRIuMAX"%%)",
+				disk_used, disk_total, disk_percentage);
 	}
 	else if (error)
-	{
-		ERR_REPORT("Could not stat $HOME for filesystem statistics.");
-	}
+		ERR_REPORT(_("Could not stat $HOME for filesystem statistics."));
 
 	return;
 }
@@ -230,93 +230,24 @@ void detect_disk(char *str)
 void detect_mem(char *str)
 {
 	FILE *mem_file;
-	long long total_mem = 0;
-	long long free_mem = 0;
-	long long used_mem = 0;
+	uintmax_t total_mem = 0, used_mem = 0, free_mem;
 
 	mem_file = popen("sysctl -n hw.memsize", "r");
-	fscanf(mem_file, "%lld", &total_mem);
+	fscanf(mem_file, "%"SCNuMAX, &total_mem);
 	pclose(mem_file);
 
 	mem_file = popen("vm_stat | head -2 | tail -1 | tr -d 'Pages free: .'", "r");
-	fscanf(mem_file, "%lld", &free_mem);
+	fscanf(mem_file, "%"SCNuMAX, &free_mem);
 	pclose(mem_file);
 
-	total_mem /= (long) MB;
+	total_mem /= (uintmax_t) MB;
 
-	free_mem *= 4096; /* 4KiB is OS X's page size */
-	free_mem /= (long) MB;
+	free_mem *= 4096; /* 4KiB is OS X's page size, only if 64-bit */
+	free_mem /= (uintmax_t) MB;
 
 	used_mem = total_mem - free_mem;
 
-	snprintf(str, MAX_STRLEN, "%lld%s / %lld%s", used_mem, "MB", total_mem, "MB");
-
-	return;
-}
-
-/*	detect_shell
-	detects the shell currently running on the computer
-	argument char *str: the char array to be filled with the shell info
-	--
-	CAVEAT: shell version detection relies on the standard versioning format for 
-	each shell. If any shell's older (or newer versions) suddenly begin to use a
-	new	scheme, the version may be displayed incorrectly.
-	--
-*/
-void detect_shell(char *str)
-{
-	FILE *shell_file;
-	char *shell_name;
-	char vers_str[MAX_STRLEN];
-
-	shell_name = getenv("SHELL");
-
-	if (shell_name == NULL)
-	{
-		if (error)
-			ERR_REPORT("Could not detect a shell.");
-
-		return;
-	}
-
-	if (STREQ(shell_name, "/bin/sh"))
-	{
-		safe_strncpy(str, "POSIX sh", MAX_STRLEN);
-	}
-	else if (strstr(shell_name, "bash"))
-	{
-		shell_file = popen("bash --version | head -1", "r");
-		fgets(vers_str, MAX_STRLEN, shell_file);
-		snprintf(str, MAX_STRLEN, "bash %.*s", 17, vers_str + 10);
-		pclose(shell_file);
-	}
-	else if (strstr(shell_name, "zsh"))
-	{
-		shell_file = popen("zsh --version", "r");
-		fgets(vers_str, MAX_STRLEN, shell_file);	
-		snprintf(str, MAX_STRLEN, "zsh %.*s", 5, vers_str + 4);
-		pclose(shell_file);
-	}
-	else if (strstr(shell_name, "csh"))
-	{
-		shell_file = popen("csh --version | head -1", "r");
-		fgets(vers_str, MAX_STRLEN, shell_file);
-		snprintf(str, MAX_STRLEN, "csh %.*s", 7, vers_str + 5);
-		pclose(shell_file);
-	}
-	else if (strstr(shell_name, "fish"))
-	{
-		shell_file = popen("fish --version", "r");
-		fgets(vers_str, MAX_STRLEN, shell_file);
-		snprintf(str, MAX_STRLEN, "fish %.*s", 13, vers_str + 6);
-		pclose(shell_file);
-	}
-	else if (strstr(shell_name, "dash") || strstr(shell_name, "ash")
-			|| strstr(shell_name, "ksh"))
-	{
-		/* i don't have a version detection system for these, yet */
-		safe_strncpy(str, shell_name, MAX_STRLEN);
-	}
+	snprintf(str, MAX_STRLEN, "%"PRIuMAX"%s / %"PRIuMAX"%s", used_mem, "MB", total_mem, "MB");
 
 	return;
 }
@@ -377,7 +308,7 @@ void detect_wm_theme(char *str, const char *wm_str)
 	OS X doesn't use GTK, so this function fills str with "Not Applicable"
 	argument char *str: the char array to be filled with any GTK info
 */
-void detect_gtk(char *str)
+void detect_gtk(char *str1, char *str2, char *str3)
 {
 	safe_strncpy(str, "Not Applicable", MAX_STRLEN);
