+++ src/prototypes.h
@@ -10,10 +10,10 @@
 
 
 /* util.c module */
-void popen_raw_shell_version(char *str1, char *str2);
+void popen_raw_shell_version(const char *str1, const char *str2);
 void remove_excess_cpu_txt(char *str1);
 char *safe_strncpy(char *destination, const char *source, size_t num);
-void split_uptime(long uptime, unsigned int *secs, unsigned int *mins, unsigned int *hrs, unsigned int *days);
+void split_uptime(long uptime, uint_least32_t *secs, uint_least32_t *mins, uint_least32_t *hrs, uint_least32_t *days);
 void take_screenshot(bool verbose, char *monitor_res);
 
 
@@ -22,7 +22,7 @@ void Init_parseR(int argc, char **argv);
 
 
 /* disp.c module */
-void process_data(char *logo[], unsigned short int num1, unsigned short int num2, char *colour);
+void process_data(const char *logo[], uint_fast16_t num, char *colour);
 void main_text_output(bool verbose);
 void main_ascii_output(char *passed_distro);
